+++ packages/terra-hyperlink/CHANGELOG.md
@@ -4,6 +4,11 @@ ChangeLog
 Unreleased
 ----------
 
+2.29.0 - (February 25, 2020)
+------------------
+### Changed
+* Update docs to use mdx syntax.
+
 2.28.0 - (February 18, 2020)
 ------------------
 ### Added